pyproject.toml
[project]
name = "sql-injection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.8"
dependencies = [
    "flask>=3.1.2",
    "mysql-connector-python>=9.5.0",
    "python-dotenv>=1.2.1",
]
